requirements.txt
streamlit
google-generativeai
pdfplumber
pandas
python-dotenv